스테이터스 레벨업 시 더미 서버에 요청 후 플레이어의 status에 적용되도록 구현 및 델리게이트를 통해 레벨업에 따른 기능이 적용되도록 구현(hp 등)

+++ Project/NyangNyang/Assets/Script/DummyServerData.cs
@@ -103,10 +103,11 @@ public class DummyServerData : MonoBehaviour
             return null;
         }
 
+        
         return enemyStatusLevelData[characterID];
     }
 
-    public static int GetUserStatusLevelFromType(int userID, StatusLevelType type)
+    public static BigInteger GetUserStatusLevelFromType(int userID, StatusLevelType type)
     {
         StatusLevelData statusLevelData = GetUserStatusLevelData(userID);
         if (statusLevelData != null)
@@ -135,7 +136,7 @@ public class DummyServerData : MonoBehaviour
         {
             GetUserStatusLevelData(userID).AddLevel(type, value);
             GetUserCurrencyData(userID).gold -= goldCost;
-
+            
             return true;
         }
